seperate filter tags to it own component
