+++ wallet/src/ui/app/components/sui-apps/index.tsx
@@ -177,4 +177,3 @@ function AppsPlayGround() {
 
 export default AppsPlayGround;
 export { default as ConnectedAppsCard } from './ConnectedAppsCard';
-export { default as AppFiltersPortal } from './filters';
